+++ src/components/Card/EndedChallengeCard.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import styled from "styled-components";
 import COLOR from "constants/color";
-import DidItIcon from "components/common/DidItIcon";
-import ChallengeFigure from "components/common/ChallengeFigure";
+import DidItIcon from "components/Challenge/DidItIcon";
+import ChallengeFigure from "components/Challenge/ChallengeFigure";
 
 const Container = styled.div`
   width: 16.3rem;
