+++ src/components/Challenge/ChallengeFigure.tsx
