+++ src/components/Challenge/DidItIcon.tsx
